+++ actix-http/tests/test_ws.rs
@@ -14,7 +14,7 @@ use actix_http::{
 use actix_http_test::test_server;
 use actix_service::{fn_factory, Service};
 use bytes::Bytes;
-use derive_more::derive::{Display, Error, From};
+use derive_more::{Display, Error, From};
 use futures_core::future::LocalBoxFuture;
 use futures_util::{SinkExt as _, StreamExt as _};