+++ actix-web/src/error/mod.rs
@@ -6,7 +6,7 @@
 //
 // See <https://github.com/rust-lang/rust/issues/83375>
 pub use actix_http::error::{ContentTypeError, DispatchError, HttpError, ParseError, PayloadError};
-use derive_more::derive::{Display, Error, From};
+use derive_more::{Display, Error, From};
 use serde_json::error::Error as JsonError;
 use serde_urlencoded::{de::Error as FormDeError, ser::Error as FormError};
 use url::ParseError as UrlParseError;