+++ actix-web/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Update `brotli` dependency to `8`.
+
 ## 4.10.2
 
 - No significant changes since `4.10.1`.