+++ actix-web/CHANGES.md
@@ -4,6 +4,7 @@
 
 - Add `HttpServer::shutdown_signal()` method.
 - Mark `HttpServer` as `#[must_use]`.
+- Re-export `mime` dependency.
 - Update `brotli` dependency to `8`.
 
 ## 4.10.2